Add combinations logic and log to the console
This commit adds utils function which combines 2 arrays to give all the possible combinations of those arrays.

If the length of value input is greater than 2, it returns null because this function can only compute 2 arrays, derived from the data which will return an array of characters for each integer in the value.